Moved all DemoItem types to Navigation.swift
