--- nimbus-ios-sample/Main/MainItem.swift
@@ -1,14 +1,0 @@
-//
-//  MainItem.swift
-//  nimbus-ios-sample
-//
-//  Created by Inder Dhir on 1/13/22.
-//
-
-enum MainItem: String, DemoItem {
-    case showAdDemo
-    case mediationPlatforms
-    case thirdPartyDemand
-    case testRender
-    case settings
-}
