--- nimbus-ios-sample/Mediation Platforms/MediationAdType.swift
@@ -1,22 +1,0 @@
-//
-//  MediationAdType.swift
-//  nimbus-ios-sample
-//
-//  Created by Victor Takai on 18/11/21.
-//
-
-import Foundation
-
-enum MediationAdType: String, DemoItem {
-    case banner, interstitial
-    case dynamicPriceBanner, dynamicPriceBannerVideo, dynamicPriceInlineVideo, dynamicPriceInterstitial
-    
-    var description: String {
-        switch self {
-        case .dynamicPriceBannerVideo:
-            return "Dynamic Price Banner + Video"
-        default:
-            return rawValue.camelCaseToWords()
-        }
-    }
-}
